Moved post notification instance to admin fragment
